requirements.txt
nose
sphinx
transitions
numpy
matplotlib
pyaudio
six
pronouncing
weighted-levenshtein
pandas
spacy
pyyaml
rospkg
catkin_pkg
scipy